moved decorateButtons to scripts.js
